Move LEDWidget to examples/platform/nrfconnect
