+++ examples/lighting-app/nrfconnect/CMakeLists.txt
@@ -16,23 +16,23 @@
 cmake_minimum_required(VERSION 3.13.1)
 
 get_filename_component(CHIP_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/third_party/connectedhomeip REALPATH)
-
-include_directories(AFTER main main/include ${CHIP_ROOT}/src/app/util)
+get_filename_component(NRFCONNECT_COMMON ${CHIP_ROOT}/examples/platform/nrfconnect REALPATH)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CHIP_ROOT}/config/nrfconnect/)
 include(nrfconnect-app)
 
 project(chip-nrf52840-lighting-example)
+target_include_directories(app PRIVATE main main/include ${NRFCONNECT_COMMON}/util/include)
 target_sources(app PRIVATE
                main/AppTask.cpp
                main/DataModelHandler.cpp
                main/LightingManager.cpp
                main/main.cpp
-               main/LEDWidget.cpp
                main/Server.cpp
                main/gen/call-command-handler.c
                main/gen/callback-stub.c
                main/gen/znet-bookkeeping.c
+               ${NRFCONNECT_COMMON}/util/LEDWidget.cpp
                ${CHIP_ROOT}/src/app/util/attribute-size.c
                ${CHIP_ROOT}/src/app/util/attribute-storage.c
                ${CHIP_ROOT}/src/app/util/attribute-table.c
