+++ examples/lighting-app/nrfconnect/main/AppTask.cpp
@@ -42,7 +42,7 @@ namespace {
 constexpr int kFactoryResetTriggerTimeout      = 3000;
 constexpr int kFactoryResetCancelWindowTimeout = 3000;
 constexpr int kAppEventQueueSize               = 10;
-constexpr int kExampleVenderID                 = 0xabcd;
+constexpr int kExampleVendorID                 = 0xabcd;
 constexpr uint8_t kButtonPushEvent             = 1;
 constexpr uint8_t kButtonReleaseEvent          = 0;
 
@@ -122,7 +122,7 @@ void AppTask::PrintQRCode() const
 
     chip::SetupPayload payload;
     payload.version       = 1;
-    payload.vendorID      = kExampleVenderID;
+    payload.vendorID      = kExampleVendorID;
     payload.productID     = 1;
     payload.setUpPINCode  = setUpPINCode;
     payload.discriminator = setUpDiscriminator;
