+++ examples/lock-app/nrfconnect/main/include/AppConfig.h
