+++ examples/platform/nrfconnect/util/include/LEDWidget.h
